move dataset test to basic
